requirements.txt
black
librosa
numpy
SpeechRecognition
openai
pydub
matplotlib
SoundFile
python-dotenv